requirements.txt
joblib==1.4.2
numpy==2.0.0
opencv_python==4.9.0.80
psycopg2_binary==2.9.9
scikit_learn==1.5.0
skimage==0.0
